fix(common): B2B-4938 PR comments
