+++ apps/storefront/src/lib/lang/locales/en.json
@@ -629,6 +629,7 @@
   "accountSettings.notification.updateEmailPassword": "Please type in your current password to update your email address.",
   "accountSettings.notification.noEdits": "You haven't made any edits",
   "accountSettings.notification.passwordNotMatch": "The current password does not match",
+  "accountSettings.notification.fieldCannotBeSaved": "This value can't be saved. Please review it and try again.",
   "accountSettings.registeredToB2b.title": "Upgrade to a business account",
   "accountSettings.registeredToB2b.description": "Access company level features like company specific pricing, buyer roles and permissions and more by completing a quick setup.",
   "accountSettings.registeredToB2b.upgrade": "Upgrade",
